+++ src/app/core/models/config.model.ts
@@ -1,0 +1,6 @@
+export interface ConfigModel {
+  id: number;
+  fullName: string;
+  email: string;
+  sessionId: string;
+}