+++ src/app/core/models/user.model.ts
@@ -1,0 +1,5 @@
+export interface UserModel {
+  id: number;
+  fullName: string;
+  email: string;
+}