+++ src/app/product/product.module.ts
@@ -1,14 +1,13 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import {ProductListComponentComponent} from "./components/product-list-component/product-list-component.component";
-import {ProductItemComponentComponent} from "./components/product-item-component/product-item-component.component";
+import {ProductListComponent} from "./components/product-list-component/product-list.component";
+import {ProductItemComponent} from "./components/product-item-component/product-item.component";
 import { ProductComponent } from './components/product/product.component';
 
 @NgModule({
   declarations: [
-    // поправьте на следующей итерации названия этих компонентов, уберите лишний суффикс Component
-    ProductListComponentComponent,
-    ProductItemComponentComponent,
+    ProductListComponent,
+    ProductItemComponent,
     ProductComponent,
   ],
   exports: [